server/requirements.txt
fastapi[all]>=0.109.0
uvicorn>=0.27.0
sqlalchemy[asyncio]>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiosqlite>=0.19.0
pydantic>=2.5.0
alembic>=1.13.0
python-dotenv>=1.0.0
argon2-cffi
pyjwt
redis
pytz
pysftp==0.2.5
aioftp==0.22.3
stripe>=11.0.0